requirements.txt
flask==2.0.1
werkzeug==2.0.3
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5
sse-starlette>=0.10.3
httpx>=0.23.0
sseclient>=0.0.27
jinja2>=3.0.0
mcp>=0.1.0
pydantic>=1.8.2,<2.0.0
typing-extensions>=4.0.0
python-dotenv>=0.19.0
aiofiles>=0.8.0 